+++ test/src/reader_test.dart
@@ -1,4 +1,4 @@
-import 'package:fpdart/src/reader.dart';
+import 'package:fpdart/fpdart.dart';
 import 'package:test/test.dart';
 
 void main() {
@@ -8,6 +8,24 @@ void main() {
       final ap = reader.map((a) => a + 1);
       final result = ap.run('abc');
       expect(result, 4);
+    });
+
+    test('map2', () {
+      final reader = Reader<String, int>((r) => r.length);
+      final reader1 = Reader<String, int>((r) => r.length * 2);
+      final ap = reader.map2<int, int>(reader1, (a, c) => a * c);
+      final result = ap.run('abc');
+      expect(result, 18);
+    });
+
+    test('map3', () {
+      final reader = Reader<String, int>((r) => r.length);
+      final reader1 = Reader<String, int>((r) => r.length * 2);
+      final reader2 = Reader<String, int>((r) => r.length * 3);
+      final ap =
+          reader.map3<int, int, int>(reader1, reader2, (a, c, d) => a * c * d);
+      final result = ap.run('ab');
+      expect(result, 48);
     });
 
     test('ap', () {
@@ -24,6 +42,58 @@ void main() {
           reader.flatMap((a) => Reader<String, int>((b) => a + b.length));
       final result = ap.run('abc');
       expect(result, 6);
+    });
+
+    test('pure', () {
+      final reader = Reader<String, int>((r) => r.length);
+      final ap = reader.pure(10);
+      final result = ap.run('abc');
+      expect(result, 10);
+    });
+
+    test('andThen', () {
+      final reader = Reader<String, int>((r) => r.length);
+      final ap =
+          reader.andThen(() => Reader<String, double>((r) => r.length / 2));
+      final result = ap.run('abc');
+      expect(result, 1.5);
+    });
+
+    test('compose', () {
+      final reader = Reader<String, int>((r) => r.length);
+      final ap = reader.compose(Reader<String, double>((r) => r.length / 2));
+      final result = ap.run('abc');
+      expect(result, 1.5);
+    });
+
+    test('local', () {
+      final reader = Reader<String, int>((r) => r.length);
+      final ap = reader.local<double>((context) => '$context');
+      final result = ap.run(7.5);
+      expect(result, 3);
+    });
+
+    test('ask', () {
+      final reader = Reader<String, int>((r) => r.length);
+      final ap = reader.ask();
+      final result = ap.run('abc');
+      expect(result, 'abc');
+    });
+
+    test('asks', () {
+      final reader = Reader<String, int>((r) => r.length);
+      final ap = reader.asks((r) => r.length * 2);
+      final result = ap.run('abc');
+      expect(result, 6);
+    });
+
+    test('flatten', () {
+      final reader = Reader<String, Reader<String, int>>(
+          (r) => Reader<String, int>((r) => r.length));
+      final ap = Reader.flatten(reader);
+      expect(ap, isA<Reader<String, int>>());
+      final result = ap.run('abc');
+      expect(result, 3);
     });
   });
 }