+++ lib/fpdart.dart
@@ -1,6 +1,7 @@
 export 'src/either.dart';
 export 'src/function.dart';
 export 'src/ilist.dart';
+export 'src/io.dart';
 export 'src/option.dart';
 export 'src/reader.dart';
 export 'src/state.dart';